Crate problem036

Problem 36 (Double-base palindromes)

Solution detail

There are fewer base-10 palindromes less than one million than there are binary palindromes. This means that we can simply iterate over the base-10 palindromes, and for each one, check if it is also a binary palindrome.

To iterate over the base-10 palindromes, we can simply iterate over all possible 'first halfs' of numbers, and create palindromes from them by reversing and gluing together.

Functions

answer
Solve the problem, returning the answer as a String

even_palindromise
Convert a number into an even-length palindrome by reversing and concatenating.

is_binary_palindrome
Check whether the given number is a binary palindrome.

main

odd_palindromise
Convert a number into an odd-length palindrome with middle digit d by reversing and concatenating.

palindromes
An iterator over all palindromes with up to the given number of digits in base 10.

solve
Find the sum of all numbers up to the given number of digits which are palindromic in base 2 and in base 10.